+++ packages/cli/src/extract-experimental/bundleSource.ts
@@ -30,6 +30,7 @@ export async function bundleSource(
     "png",
     "css",
     "sass",
+    "scss",
     "less",
     "jpg",
   ]